Implemented histogram read
git-svn-id: svn+ssh://savannah.psi.ch/afs/psi.ch/project/meg/svn/rome/trunk/rome@901 306b0a18-5003-0410-92f3-cc1c673a7936

+++ examples/tree_write_read/src/tasks/TWRTFillTree.cpp
@@ -32,7 +32,7 @@ void TWRTFillTree::BeginOfRun()
 
 void TWRTFillTree::Event()
 {
-   gAnalyzer->GetData()->SetValue(gAnalyzer->GetCurrentEventNumber());
+   gAnalyzer->GetData()->SetValue((int)gAnalyzer->GetCurrentEventNumber());
 }
 
 void TWRTFillTree::EndOfRun()
